requirements.txt
selenium
beautifulsoup4
pandas
streamlit
google-generativeai
PyPDF2
scikit-learn
